Name the internals files for what they hold, and give constants a home
`internals` said only "the stuff that is not the other stuff" — the same
non-answer `-core` gave in a package name. All three are gone, replaced by
names that describe their contents. They did not need the SAME name,
because they did not hold the same kind of thing:

  graph/internals.ts   -> constants.ts        (it was four frozen empties)
  commands/internals.ts -> results.ts + queries.ts
  patches/internals.ts  -> constants.ts + results.ts + predicates.ts
                           + splicing.ts

`patches/internals.ts` had drifted to FOUR concerns — two constants, a
rejection constructor, four predicates, and a hundred lines of list
splicing. That is what a name meaning "miscellaneous" collects, and the
argument for not having one.

CONSTANTS, measured rather than assumed. Every module-level constant was
checked for whether it is used INSIDE the file that declares it or only
elsewhere:

  moved   NO_IDS, NO_OWNERS, NO_PLACEMENTS, NO_DEAD_REVS   (graph, 3-4 users)
  moved   EMPTY_IDS, VERIFY_OK                             (patches, 2-5)
  moved   FOLD_CACHE_HEADROOM, DEFAULT_INTERACTIVE_NODE_BUDGET,
          SHADOW_REFOLD_BUDGET                             (engine)
  moved   DEFAULT_MAX_NODES                                (serialize)
  STAYED  DEFAULT_FOLD_CACHE_LIMIT, CERTAINTY_RANK,
          DESCRIBE_LIMIT, DEV_CHECK_BUDGET

The four that stayed are read by the file that declares them, and each
carries the measurement that chose its value. Moving those away from their
only caller would have cost the context that explains them and bought
nothing. `SHADOW_REFOLD_BUDGET` and `DEFAULT_MAX_NODES` went the other way
for the same reason read backwards: neither is referenced by the module
that declared it, so neither was ever "beside its user".

`engine/defaults.ts` keeps the default IMPLEMENTATIONS — `defaultMintId`,
`noop`, `sameIds` — which is what the name always meant.

All eight folders stay value-acyclic; `types` keeps its one type-only
cycle.

Gate: 7/7 packages typecheck, noUnusedLocals clean, lint:packages and app
lint 0 errors, app tsc clean, 1555 unit tests across the same 85 files.
